+++ examples/svelte-example/src/routes/+layout.ts
@@ -1,0 +1,6 @@
+import '@uppy/core/dist/style.min.css'
+import '@uppy/dashboard/dist/style.min.css'
+import '@uppy/drag-drop/dist/style.min.css'
+import '@uppy/progress-bar/dist/style.min.css'
+
+export const ssr = false