+++ packages/@uppy/locales/src/bg_BG.ts
@@ -166,11 +166,4 @@ bg_BG.strings = {
   },
 }
 
-// TODO: remove this in the next major?
-// @ts-expect-error Uppy can be a global in legacy bundle
-if (typeof Uppy !== 'undefined') {
-  // @ts-expect-error Uppy can be a global in legacy bundle
-  globalThis.Uppy.locales.bg_BG = bg_BG
-}
-
 export default bg_BG